scraper/requirements.txt
aiohttp
asyncio
GitPython
python-dotenv
aiofiles